+++ integration_test/common_tests/test_dump_tag.dart
@@ -1,0 +1,60 @@
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:miziptools/extensions/uint8list_extensions.dart';
+import 'package:miziptools/main.dart';
+import 'package:miziptools/widgets/dump/dump_tag.dart' show DumpTag;
+import 'package:path_provider/path_provider.dart';
+import '../mock/mock_nfc_adapter.dart';
+import '../mock/mock_nfc_tag.dart';
+import 'consts.dart';
+
+Future<void> testDumpTagRemovedTag(WidgetTester tester, MockNfcTag mockTag) async{
+  final mockAdapter = MockNfcAdapter();
+  mockAdapter.setTag(mockTag);
+  await commonDumpTagExec(tester, mockAdapter, "Communication error", disconnectTag: true);
+}
+
+Future<void> testDumpTagWrongKey(WidgetTester tester, MockNfcTag mockTag) async{
+  final mockAdapter = MockNfcAdapter();
+  mockTag.setDenyAuthList([3]);
+  mockAdapter.setTag(mockTag);
+  await commonDumpTagExec(tester, mockAdapter, "Incorrect keys");
+}
+
+Future<void> testDumpTagSuccess(WidgetTester tester, MockNfcTag mockTag) async{
+  final mockAdapter = MockNfcAdapter();
+  mockAdapter.setTag(mockTag);
+  await commonDumpTagExec(tester, mockAdapter, "Dump done file : ${mockAdapter.currentTag!.getUid().toUpperCase()}.dump");
+}
+
+Future<void> commonDumpTagExec(WidgetTester tester, MockNfcAdapter mockAdapter, String expectedSnackBarMessage, {bool disconnectTag = false}) async{
+  final dir = await getExternalStorageDirectory();
+  await tester.pumpWidget(App(nfcAdapter: mockAdapter, dataDir: dir!,));
+  await tester.tap(find.widgetWithText(Tab, "Dumps"));
+  await tester.pumpAndSettle(delay);
+  if (disconnectTag){
+    mockAdapter.setTagRemoved(true);
+  }
+  await tester.tap(find.widgetWithText(DumpTag, "Dump Tag"));
+  await tester.pumpAndSettle(delay);
+  // Second delay for second SnackBar
+  await tester.pumpAndSettle(delay);
+  expect(find.widgetWithText(SnackBar, expectedSnackBarMessage), findsOneWidget);
+      
+  // Only if the dump was supposed to be successful
+    if (mockAdapter.currentTag!.denyAuthList.isEmpty && !disconnectTag){
+    // File exists
+    final file = dir.listSync().where((f) => f.path.split("/").last != "uid_save").last;
+    expect(file.path.split("/").last, equals("${mockAdapter.currentTag!.getUid().toUpperCase()}.dump"));
+
+    // Content is OK
+    final expectedContent = mockAdapter.currentTag!.data.map((block) => block.toHexString().toUpperCase()).join("\n");
+    final fileContent = File(file.path).readAsStringSync();
+    expect(fileContent, equals(expectedContent));
+
+    // Cleanup
+    File(file.path).deleteSync();
+  }
+}